More portability improvements (#3173)
* Size more struct allocations

* Size the fighter object pools from their element types

* Type the stage ground DAT's opaque fields

* Type ftData::x20 as ftData_x20

* Type Fighter_804D652C and FtSFX::x20

* Make debug collision drawing use real types

* Cast lbArq_80014BD0's dest pointer through uintptr_t

+++ src/melee/lb/lbarq.c
@@ -131,7 +131,7 @@ void lbArq_80014BD0(unsigned int source, void* dest, size_t length,
 
     rp_tmp = rp;
     source_tmp = source;
-    ARQPostRequest(&rp->arq, (u32) rp_tmp, 1, 0, source_tmp, (u32) dest,
+    ARQPostRequest(&rp->arq, (u32) rp_tmp, 1, 0, source_tmp, (uintptr_t) dest,
                    length, (ARQCallback) lbArq_80014AC4);
 
     if (rp->callback == NULL) {
